Implemented changes/issues reported by Richard.
When I modify or add a threshold – it seems to update the underlying DB
but does not add or change the dotted threshold lines on the graph
KK: The threshold updated would be applicable for future readings. The
old values will have old threshold limits. Hence it will not update the
graph for old readings.
RR: Agreed but could you check for example the simple inclusive
threshold for pulse for patient d9bc2478-062e-4b87-9060-4984f26b74be
which has an effective date of 2016-11-04 16:21:34.589. I added new data
this morning and no threshold lines have appeared

Could the Temperature thresholds only have 1 decimal place please?
KK: All the div’s are generic. If we want to change this, it would be
applicable to all the attributes. Should I change this for  all the
attributes then?
RR: yes 1dp for now please

When a new value is being input it’s too big to be shown whole in the
box  -could you change it so that it doesn’t scroll out of the box?
KK: Do you want the font size should be small when the user is trying to
edit the values?
RR: Yes that would help